add `--tag` on diff and diff-all
