+++ projects/optic/src/commands/diff/diff.ts
@@ -67,6 +67,10 @@ export const registerDiff = (cli: Command, config: OpticCliConfig) => {
       '--standard <standard>',
       'run comparison with a locally defined standard, if not set, looks for the standard on the [x-optic-standard] key on the spec, and then the optic.dev.yml file.'
     )
+    .option(
+      '--tag <tag>',
+      'Adds additional tags to the HEAD spec. Should be used in conjunction with `--upload`'
+    )
     .addOption(new Option('--ruleset <ruleset>', '').hideHelp())
     .option('--check', 'enable checks', false)
     .option('--upload', 'upload run to cloud', false)
@@ -202,6 +206,7 @@ type DiffActionOptions = {
   json: boolean;
   standard?: string;
   ruleset?: string;
+  tag?: string;
 };
 
 const getDiffAction =
@@ -269,7 +274,8 @@ const getDiffAction =
           to: headParseResult,
         },
         diffResult.specResults,
-        config
+        config,
+        options
       );
     }
